+++ CHANGELOG.md
@@ -24,8 +24,8 @@ Breaking changes
 - @iov/bns-governance: Remove `CommitteeIds` type as well as `committeeIds` and
   `guaranteeFundEscrowIds` constants. These are now the responsibility of the
   consumer.
-- @iov/bns-governance: Remove `TreasurySend` proposal type and
-  `TreasurySendOptions` type.
+- @iov/bns-governance: Remove `TreasurySend` proposal type,
+  `TreasurySendOptions` type and `GovernorOptions.treasuryAddress`.
 - @iov/bns-governance: Make `guaranteeFundEscrowId` `Governor` option numeric.
 - @iov/encoding: Remove deprecated `EnglishMnemonic.asString` use `.toString`
   instead.